Build abstraction around pools
