+++ priv/cert/selfsigned.pem
@@ -1,0 +1,22 @@
+-----BEGIN CERTIFICATE-----
+MIIDZDCCAkygAwIBAgIITz/jLOEeziMwDQYJKoZIhvcNAQELBQAwJzElMCMGA1UE
+AwwcU2VsZi1zaWduZWQgdGVzdCBjZXJ0aWZpY2F0ZTAeFw0yMDA0MjUxNzQ2MTVa
+Fw0yMTA1MjUxNzUxMTVaMCcxJTAjBgNVBAMMHFNlbGYtc2lnbmVkIHRlc3QgY2Vy
+dGlmaWNhdGUwggEiMA0GCSqGSIb3DQEBAQUAA4IBDwAwggEKAoIBAQD7FyQFJZz5
+OwWuWz+wfhJzST4KtdM0XbTNVXVoXsq9cX/bbbOxH2igtwVYtpayqc4GLR4tHxIR
+478D89043RNBnX79zGSFpApgsIj6/uNJroLZubswUcWVucy2mCFzaLWDvHHrfXa0
+Uhk5czmthPZPcbKgJygvV/Z+cWxXaMI8VbcXuZjNyIGRY8oEIee7fO84Loufz2XW
+r0F/QoJjEZgl4sY02zkegm7QJGCTOIcxgziEV8h8XG4HTNvixWBruUqyG53o/Ei+
+SBJ9lijJwEfnhK3Jlu1Lcx/PTkmE0DahlbR5eUlKBS6UE6vIUL/C4r+Dw0/0/vZi
+ebTaJ8Phdmf7AgMBAAGjgZMwgZAwCQYDVR0TBAIwADAOBgNVHQ8BAf8EBAMCBaAw
+HQYDVR0lBBYwFAYIKwYBBQUHAwEGCCsGAQUFBwMCMB0GA1UdDgQWBBSqqsioL/Tb
+JRfkB+BRsWvVqKBdWTAfBgNVHSMEGDAWgBSqqsioL/TbJRfkB+BRsWvVqKBdWTAU
+BgNVHREEDTALgglsb2NhbGhvc3QwDQYJKoZIhvcNAQELBQADggEBAObzW+gguEpV
+zpqWsEGz0hKNPlxdv4iTwm/EVB2S/Pm9PdlT2SgO5rr+Beity7m9mj4NtKpU7tIq
+c9KNUPYG+bxO32U7TFEvCXonicX2Fhzoflr02fYn78rd28A2+jn3cK6uzhPV5wA5
+gYEbsMKt9jXiaq1qmXvlphuddMkbFYKRTC4SzRgCFwtXRsvIVBzXhTP6tkxHYjMR
+erO7xV0Hhbg15iNnkrP9kx1kOLWwDDccX2S2Cpdwx7o9AuU8za7plCsgilDHSMzu
+BOb6iswxz4SzhHCJWTd6JKI4CLn3dBaXCqKkZKlRcyXBCsWtYNCYXXvVltNX92ED
+Gg0gWPR3hbI=
+-----END CERTIFICATE-----
+
