+++ priv/cert/selfsigned_key.pem
@@ -1,0 +1,28 @@
+-----BEGIN RSA PRIVATE KEY-----
+MIIEowIBAAKCAQEA+xckBSWc+TsFrls/sH4Sc0k+CrXTNF20zVV1aF7KvXF/222z
+sR9ooLcFWLaWsqnOBi0eLR8SEeO/A/PdON0TQZ1+/cxkhaQKYLCI+v7jSa6C2bm7
+MFHFlbnMtpghc2i1g7xx6312tFIZOXM5rYT2T3GyoCcoL1f2fnFsV2jCPFW3F7mY
+zciBkWPKBCHnu3zvOC6Ln89l1q9Bf0KCYxGYJeLGNNs5HoJu0CRgkziHMYM4hFfI
+fFxuB0zb4sVga7lKshud6PxIvkgSfZYoycBH54StyZbtS3Mfz05JhNA2oZW0eXlJ
+SgUulBOryFC/wuK/g8NP9P72Ynm02ifD4XZn+wIDAQABAoIBACgn1kHgdw9AyjLN
+Mw1VzDuVvg9WbtWUGuMvxJwRdNkBOtJr4+3yuPfi7T73zHf8Qg+ozYenzWsBAj5u
+lLWQXZED0BhkxLq+tk4qTW3er5jMCO8rDe23ViMQ2NVm1Qqw9z5/N7TSltZVvuT6
+MYI3nnCsHUYRnRKdPbWRDCOahF3DgnD+q5MFDdrVasdDuZrMfoQ+7tMbkQYSNOkn
+ykY86NYypRuhbgnSNra3F7GO3YiEXqVq7rzbN2r7ymUNVdYx7dh6rtTXgpb+7/8h
+MPMKYmHl95PIPSnQQxrOi3CarysvXbm0IodRgUDhpWZkzm0KNoRD5a3lIRjGkPQY
+os/7JVECgYEA/2E2YdLH4GkESsFIcQiXAMNAnIrU5J5H5igNEjoT1vO3kzNu/+a7
+5c2Hzmg4HHIuu9GyCYvG7DR7aj3YlzaDBzSgGmL13Rd+AaEeoM+2TMJHfpouwzqS
+gy/5/gWqGTSauCagoSAnzKbUcLRgaUGzLUZp6Kh85MzMXexpU2tkYoMCgYEA+7NC
+47LDDolJD1QU0en5dsFkwrNRNfmsoID7qYJ7AUjyFfUTLXBWiGa8KsjrxocbD0EB
+7C4pMwXpFAbHWE1Gxiede0eJhjB2BOQoAD2Q/jEToCIZhERILSsyPX9XXJFin61k
+zX+e45cmmnUGFKh+pXNX13bXb05pqgHCFLSNCykCgYBllnAhnc1w0ZWpfxWoETiw
+TvtmyjtB7ky/m3MHqx61RSHjZnRjS4+3LHk3fHHFrw40HIhsD7RDxeTKMGv2SvL4
+bNC7bl5LYsr3LvaWFcp44le6Xu73LleV5AHYMLynZiud0Y/YSIhMepjLdJVdiYsT
+GY2LxWpZQbaSLjb/vhqirQKBgFYl6ewDQE3WYUI01hduzKp/0NvvAgnGVfZtjzVa
+an1rGSi4xAKWY5e2Cgtez4wYrANRjbXnqsRC7p6jH1GyyVmDL5yFRKXAKnTMHoDO
+kbTe2UfCxBcIQ+fp0/1Ge4E8ZZQfhdPsLVbaYiUVjqo9bXN1gghGLcB0ByhgB6yM
+i4XZAoGBAPQYDYid4xK1snD7wWWi835p+bN5cezuG9hlRg1QdjBxsdKyfxRafREv
+WszS3dcC8ADsV2ZSAyaiTetaBkDTnzfTsvdCzbMQVBWOterampHMgmt4PGTqoHOJ
+MoNYdBGZUAGs5/kdyVkWktwuwJAYFh0kf95oGrHsf7jpOZyxIWgX
+-----END RSA PRIVATE KEY-----
+
